requirements.txt
aioredis==1.2.0
asgiref==2.3.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.2.1
Automat==0.7.0
bcrypt==3.1.4
boto3==1.7.45
botocore==1.10.84
cffi==1.12.2
channels==2.1.7
channels-redis==2.3.3
constantly==15.1.0
daphne==2.2.5
dj-database-url==0.5.0
Django==1.11.21
django-storages==1.6.6
djangorestframework==3.9.1
docutils==0.14
gunicorn==19.9.0
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
jmespath==0.9.4
msgpack==0.6.1
Pillow==5.4.1
psycopg2==2.7.6.1
pycparser==2.19
PyHamcrest==1.9.0
python-dateutil==2.8.0
pytz==2018.9
s3transfer==0.1.13
six==1.12.0
Twisted==19.2.1
txaio==18.8.1
whitenoise==3.3.1
zope.interface==4.6.0
